http: share path-param substitution across protocols; key OAuth cache by full config
Pre-release review of dev. Two fixes cubic found in the streaming
protocols also applied to the plain HTTP protocol, where they were
pre-existing:

- Path parameters: a single buildUrlWithPathParams helper (new _url.ts)
  now backs all three protocols. Every occurrence of a parameter is
  substituted (the HTTP protocol replaced only the first, then threw
  "Missing required path parameter" for the repeat), and the `${param}`
  form the README documents works everywhere.
- The HTTP protocol's OAuth2 token cache is keyed by token URL, client
  id, secret and scope instead of client_id alone, matching the SSE and
  Streamable HTTP protocols.
- Removed a stale duplicate comment in _normalizeToolError.
